Some TS applicative tests
