+++ Fable.sln
@@ -24,37 +24,39 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 		build.fsx = build.fsx
 	EndProjectSection
 EndProject
-Project("{F2A71F9B-5D33-465A-A702-920D77279786}") = "Rust.AST", "src\Fable.Transforms\Rust\AST\Rust.AST.fsproj", "{218FBF6A-D01C-4F28-9146-50706C22D5AC}"
+Project("{6EC3EE1D-3C4E-46DD-8F32-0CC8E7565705}") = "Rust.AST", "src\Fable.Transforms\Rust\AST\Rust.AST.fsproj", "{218FBF6A-D01C-4F28-9146-50706C22D5AC}"
 EndProject
-Project("{F2A71F9B-5D33-465A-A702-920D77279786}") = "Fable.Tests.External", "tests_external\Fable.Tests.External.fsproj", "{78507119-D0CA-4C25-AE2B-361BAE8B15EF}"
+Project("{6EC3EE1D-3C4E-46DD-8F32-0CC8E7565705}") = "Fable.Tests.External", "tests_external\Fable.Tests.External.fsproj", "{78507119-D0CA-4C25-AE2B-361BAE8B15EF}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "tests", "tests", "{CF0A8EC3-841F-4A54-B9FC-8D174CCD4A90}"
 EndProject
-Project("{F2A71F9B-5D33-465A-A702-920D77279786}") = "Fable.Tests.React", "tests\React\Fable.Tests.React.fsproj", "{11256F5D-E00D-40B7-A2ED-EA24AC33E2B6}"
+Project("{6EC3EE1D-3C4E-46DD-8F32-0CC8E7565705}") = "Fable.Tests.React", "tests\React\Fable.Tests.React.fsproj", "{11256F5D-E00D-40B7-A2ED-EA24AC33E2B6}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Js", "Js", "{DA29278E-3808-42DE-8333-964F129F295D}"
 EndProject
-Project("{F2A71F9B-5D33-465A-A702-920D77279786}") = "Fable.Tests.DllRef", "tests\Js\DllRef\Fable.Tests.DllRef.fsproj", "{1A7D4766-35B7-4932-A9D8-62CCF34D8959}"
+Project("{6EC3EE1D-3C4E-46DD-8F32-0CC8E7565705}") = "Fable.Tests.DllRef", "tests\Js\DllRef\Fable.Tests.DllRef.fsproj", "{1A7D4766-35B7-4932-A9D8-62CCF34D8959}"
 EndProject
-Project("{F2A71F9B-5D33-465A-A702-920D77279786}") = "Fable.Tests", "tests\Js\Main\Fable.Tests.fsproj", "{CC558878-A44C-4C12-9C3C-0E0522C68965}"
+Project("{6EC3EE1D-3C4E-46DD-8F32-0CC8E7565705}") = "Fable.Tests", "tests\Js\Main\Fable.Tests.fsproj", "{CC558878-A44C-4C12-9C3C-0E0522C68965}"
 EndProject
-Project("{F2A71F9B-5D33-465A-A702-920D77279786}") = "Fable.Tests.Adaptive", "tests\Js\Adaptive\Fable.Tests.Adaptive.fsproj", "{56276592-4CF4-40B2-9202-C7A637EFBDEE}"
+Project("{6EC3EE1D-3C4E-46DD-8F32-0CC8E7565705}") = "Fable.Tests.Adaptive", "tests\Js\Adaptive\Fable.Tests.Adaptive.fsproj", "{56276592-4CF4-40B2-9202-C7A637EFBDEE}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Integration", "Integration", "{3FE69E51-7170-4B63-A1CD-95F6E1233588}"
 EndProject
-Project("{F2A71F9B-5D33-465A-A702-920D77279786}") = "Fable.Tests.Integration", "tests\Integration\Integration\Fable.Tests.Integration.fsproj", "{DCB18069-33B1-4D39-A079-61A23C6BCA24}"
+Project("{6EC3EE1D-3C4E-46DD-8F32-0CC8E7565705}") = "Fable.Tests.Integration", "tests\Integration\Integration\Fable.Tests.Integration.fsproj", "{DCB18069-33B1-4D39-A079-61A23C6BCA24}"
 EndProject
-Project("{F2A71F9B-5D33-465A-A702-920D77279786}") = "Fable.Tests.Compiler", "tests\Integration\Compiler\Fable.Tests.Compiler.fsproj", "{475E8FB8-6BC4-4F2B-8081-4D5626181A6A}"
+Project("{6EC3EE1D-3C4E-46DD-8F32-0CC8E7565705}") = "Fable.Tests.Compiler", "tests\Integration\Compiler\Fable.Tests.Compiler.fsproj", "{475E8FB8-6BC4-4F2B-8081-4D5626181A6A}"
 EndProject
-Project("{F2A71F9B-5D33-465A-A702-920D77279786}") = "Fable.Tests.Rust", "tests\Rust\Fable.Tests.Rust.fsproj", "{70D726EE-D492-4561-9C86-9C1CAAAC2ABD}"
+Project("{6EC3EE1D-3C4E-46DD-8F32-0CC8E7565705}") = "Fable.Tests.Rust", "tests\Rust\Fable.Tests.Rust.fsproj", "{70D726EE-D492-4561-9C86-9C1CAAAC2ABD}"
 EndProject
-Project("{F2A71F9B-5D33-465A-A702-920D77279786}") = "Fable.Tests.Python", "tests\Python\Fable.Tests.Python.fsproj", "{2C27ECA6-6502-4A47-8D19-6102840B911B}"
+Project("{6EC3EE1D-3C4E-46DD-8F32-0CC8E7565705}") = "Fable.Tests.Python", "tests\Python\Fable.Tests.Python.fsproj", "{2C27ECA6-6502-4A47-8D19-6102840B911B}"
 EndProject
-Project("{F2A71F9B-5D33-465A-A702-920D77279786}") = "Fable.Library.Dart", "src\fable-library-dart\Fable.Library.Dart.fsproj", "{505AEE8D-2BB6-48D1-ACD2-FB262DB365C4}"
+Project("{6EC3EE1D-3C4E-46DD-8F32-0CC8E7565705}") = "Fable.Library.Dart", "src\fable-library-dart\Fable.Library.Dart.fsproj", "{505AEE8D-2BB6-48D1-ACD2-FB262DB365C4}"
 EndProject
-Project("{F2A71F9B-5D33-465A-A702-920D77279786}") = "Quicktest.Dart", "src\quicktest-dart\Quicktest.Dart.fsproj", "{2971A69B-4A5B-4841-920F-38A5D11F7881}"
+Project("{6EC3EE1D-3C4E-46DD-8F32-0CC8E7565705}") = "Quicktest.Dart", "src\quicktest-dart\Quicktest.Dart.fsproj", "{2971A69B-4A5B-4841-920F-38A5D11F7881}"
 EndProject
-Project("{F2A71F9B-5D33-465A-A702-920D77279786}") = "Fable.Tests.Dart", "tests\Dart\src\Fable.Tests.Dart.fsproj", "{40C7000A-6BC2-42B7-B7E2-3AAC60DAC06D}"
+Project("{6EC3EE1D-3C4E-46DD-8F32-0CC8E7565705}") = "Fable.Tests.Dart", "tests\Dart\src\Fable.Tests.Dart.fsproj", "{40C7000A-6BC2-42B7-B7E2-3AAC60DAC06D}"
+EndProject
+Project("{6EC3EE1D-3C4E-46DD-8F32-0CC8E7565705}") = "Fable.Tests.TypeScript", "tests\TypeScript\Fable.Tests.TypeScript.fsproj", "{F9134F40-C6CD-4368-A28E-33E5102FD4AB}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -142,6 +144,10 @@ Global
 		{40C7000A-6BC2-42B7-B7E2-3AAC60DAC06D}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{40C7000A-6BC2-42B7-B7E2-3AAC60DAC06D}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{40C7000A-6BC2-42B7-B7E2-3AAC60DAC06D}.Release|Any CPU.Build.0 = Release|Any CPU
+		{F9134F40-C6CD-4368-A28E-33E5102FD4AB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{F9134F40-C6CD-4368-A28E-33E5102FD4AB}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{F9134F40-C6CD-4368-A28E-33E5102FD4AB}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{F9134F40-C6CD-4368-A28E-33E5102FD4AB}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -168,6 +174,7 @@ Global
 		{505AEE8D-2BB6-48D1-ACD2-FB262DB365C4} = {C8CB96CF-68A8-4083-A0F8-319275CF8097}
 		{2971A69B-4A5B-4841-920F-38A5D11F7881} = {C8CB96CF-68A8-4083-A0F8-319275CF8097}
 		{40C7000A-6BC2-42B7-B7E2-3AAC60DAC06D} = {CF0A8EC3-841F-4A54-B9FC-8D174CCD4A90}
+		{F9134F40-C6CD-4368-A28E-33E5102FD4AB} = {CF0A8EC3-841F-4A54-B9FC-8D174CCD4A90}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {58DF9285-8523-4EAC-B598-BE5B02A76A00}
