+++ src/fable-library/Option.ts
@@ -1,4 +1,5 @@
-import { compare, equals, structuralHash } from "./Util.js";
+import { Nullable, Option, Some, value } from "./Util.js";
+export { Nullable, Option, Some, value };
 
 // Options are erased in runtime by Fable, but we have
 // the `Some` type below to wrap values that would evaluate
@@ -12,58 +13,9 @@ import { compare, equals, structuralHash } from "./Util.js";
 // Note: We use non-strict null check for backwards compatibility with
 // code that use F# options to represent values that could be null in JS
 
-export type Nullable<T> = T | undefined;
-
-export type Option<T> = T | Some<T> | undefined;
-
-// Using a class here for better compatibility with TS files importing Some
-export class Some<T> {
-  public value: T;
-
-  constructor(value: T) {
-    this.value = value;
-  }
-
-  public toJSON() {
-    return this.value;
-  }
-
-  // Don't add "Some" for consistency with erased options
-  public toString() {
-    return String(this.value);
-  }
-
-  public GetHashCode() {
-    return structuralHash(this.value);
-  }
-
-  public Equals(other: Option<T>): boolean {
-    if (other == null) {
-      return false;
-    } else {
-      return equals(this.value, other instanceof Some ? other.value : other);
-    }
-  }
-
-  public CompareTo(other: Option<T>) {
-    if (other == null) {
-      return 1;
-    } else {
-      return compare(this.value, other instanceof Some ? other.value : other);
-    }
-  }
-}
 
 export function some<T>(x: T): Option<T> {
   return x == null || x instanceof Some ? new Some(x) : x;
-}
-
-export function value<T>(x: Option<T>) {
-  if (x == null) {
-    throw new Error("Option has no value");
-  } else {
-    return x instanceof Some ? x.value : x;
-  }
 }
 
 export function ofNullable<T>(x: T | null): Option<T> {
