+++ tests/TypeScript/Fable.Tests.TypeScript.fsproj
@@ -10,6 +10,7 @@
     <Compile Include="../Js/Main/Util/Util.Testing.fs" />
     <Compile Include="../Js/Main/Util/Util.fs" />
     <Compile Include="../Js/Main/Util/UtilTests.fs" />
+    <Compile Include="../Js/Main/ApplicativeTests.fs" />
     <Compile Include="../Js/Main/ArithmeticTests.fs" />
     <Compile Include="../Js/Main/ArrayTests.fs" />
     <Compile Include="../Js/Main/ListTests.fs" />
